deprecated/automation/integrated_automation/requirements.txt
# 통합 클라우드 과정 자동화 시스템 의존성

# 기본 Python 라이브러리
pyyaml>=6.0
requests>=2.28.0
psutil>=5.9.0

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# Google Cloud SDK
google-cloud-storage>=2.7.0
google-cloud-compute>=1.8.0

# Docker SDK
docker>=6.0.0

# Kubernetes SDK
kubernetes>=24.2.0

# GitHub API
PyGithub>=1.58.0

# Terraform (Python wrapper)
python-terraform>=0.10.0

# 로깅 및 모니터링
colorlog>=6.7.0
rich>=13.0.0

# 테스트
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# 개발 도구
black>=22.0.0
flake8>=6.0.0
mypy>=1.0.0

# 문서화
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
